\name{fastaitvar}
\alias{fastaitvar}
\title{compute aitchison variation of matrix cols quickly}
\usage{
fastaitvar(x)
}
\arguments{
  \item{x}{NumericMatrix with data}
}
\value{
symmetric matrix with pairwise Aitchison variations
}
\description{
compute aitchison variation of matrix cols quickly
}

