\name{fixed_mean_vector}
\alias{fixed_mean_vector}
\title{Draw mean vector from a vector of fixed means & pick the abundance values for a
fixed number of bugs with higher means that also matches the desired n_eff}
\usage{
fixed_mean_vector(D, neff, val = 2, numhigher = ceiling(D/50))
}
\description{
Draw mean vector from a vector of fixed means & pick the
abundance values for a fixed number of bugs with higher
means that also matches the desired n_eff
}

