\name{get_comm_params}
\alias{get_comm_params}
\title{Get the parameters for the OTUs (rows) of each community}
\usage{
get_comm_params(comm, distr, ...)
}
\arguments{
  \item{comm}{community: matrix of counts}

  \item{distr}{distribution to fit (see fitdistr)}
}
\value{
list of parameters
}
\description{
Get the parameters for the OTUs (rows) of each community
}

