\name{graph2prec}
\alias{graph2prec}
\title{Convert a symmetric graph (extension of R matrix class)}
\usage{
graph2prec(Graph, u = 0.1, v = 0.3, method)
}
\arguments{
  \item{u}{some parameter that controls Prec values}

  \item{v}{some other parameter that controls Prec values}
}
\description{
Has internal rules for converting various graph topologies
into the associated adjancency and, therefore, precision
matrix
}

