\name{log_norm_mean_vector}
\alias{log_norm_mean_vector}
\title{Draw mean vector from a log-normal vector & find the logsd parameter that
matches the desired n_eff}
\usage{
log_norm_mean_vector(D, logmean = 1, neff)
}
\description{
Draw mean vector from a log-normal vector & find the logsd
parameter that matches the desired n_eff
}

