\name{proportionality_filter}
\alias{proportionality_filter}
\title{compute p-vals of proportionality (two ways) via aitchison variance}
\usage{
proportionality_filter(data, normfun = function(x) x, nboots = 1333,
  mar = 2, n.core = 1, ...)
}
\arguments{
  \item{data}{OTUs/features in columns, samples are rows}
}
\description{
compute p-vals of proportionality (two ways) via aitchison
variance
}

