\name{sparseSigma_JF}
\alias{sparseSigma_JF}
\title{Jonathan Friedman's described method of computing a correlation matrx}
\usage{
sparseSigma_JF(D, p, var = 1, center = 0)
}
\arguments{
  \item{D}{Dimension}

  \item{p}{probability that any otu pairs are perfectly
  correlated}

  \item{var}{individual variance (same for all OTUs)}

  \item{center}{average correlation (default 0.0)}
}
\description{
Jonathan Friedman's described method of computing a
correlation matrx
}

